Refactor _global and getObjectDiff to own modules
