+++ src/helpers/promise.js
@@ -2,6 +2,7 @@
  * Copyright (c) 2014-2017 David Fahlander
  * Apache License Version 2.0, January 2004, http://www.apache.org/licenses/LICENSE-2.0
  */
+import { _global } from '../globals/global';
 import {tryCatch, props, setProp, _global,
     getPropertyDescriptor, getArrayOf, extend, getProto} from '../functions/utils';
 import {nop, callBoth, mirror} from '../functions/chaining-functions';
